python/bc-crypto/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "bc-crypto"
version = "0.14.0"
description = "Blockchain Commons cryptographic primitives"
requires-python = ">=3.10"
license = "BSD-2-Clause-Patent"
keywords = ["cryptography", "blockchain", "bitcoin"]
dependencies = [
  "bc-rand==0.5.0",
  "cryptography>=42.0.0",
  "btclib>=2023.7.12",
  "argon2-cffi>=23.1.0",
]

[project.optional-dependencies]
dev = ["pytest"]

[tool.hatch.build.targets.wheel]
packages = ["src/bc_crypto"]
